[mod,del] move hooks out of individual folders
