--- src/hooks/useBreakpointObserver/index.ts
@@ -1,1 +1,0 @@
-export { useBreakpointObserver } from "./useBreakpointObserver";
