+++ src/hooks/useCategoryValidation.tsx
