--- src/hooks/useCategoryValidation/index.ts
@@ -1,1 +1,0 @@
-export { useCategoryValidation } from "./useCategoryValidation";
