--- src/hooks/useContextMenu/index.ts
@@ -1,1 +1,0 @@
-export { useContextMenu } from "./useContextMenu";
