--- src/hooks/useDebounce/index.ts
@@ -1,1 +1,0 @@
-export { useDebounce } from "./useDebounce";
