--- src/hooks/useDialog/index.ts
@@ -1,1 +1,0 @@
-export { useDialog, useDialogHotkey } from "./useDialog";
