--- src/hooks/useDndFileDrop/index.ts
@@ -1,1 +1,0 @@
-export { useDndFileDrop } from "./useDndFileDrop";
