--- src/hooks/useErrorHandler/index.ts
@@ -1,1 +1,0 @@
-export { useErrorHandler } from "./useErrorHandler";
