--- src/hooks/useHotkeys/index.ts
@@ -1,1 +1,0 @@
-export { useHotkeys } from "./useHotkeys";
