--- src/hooks/useLocalGlobalState/index.ts
@@ -1,1 +1,0 @@
-export { useLocalGlobalState } from "./useLocalGlobalState";
