--- src/hooks/useMenu/index.ts
@@ -1,1 +1,0 @@
-export { useMenu } from "./useMenu";
