--- src/hooks/useUnloadConfirmation/index.ts
@@ -1,1 +1,0 @@
-export { useUnloadConfirmation } from "./useUnloadConfirmation";
